refactored a little and fixed bug
fixes: #2895
