+++ src/components/client/ClientSelect.tsx
@@ -1,0 +1,102 @@
+import React, { useState } from "react";
+import { useTranslation } from "react-i18next";
+import { Controller, useFormContext } from "react-hook-form";
+import {
+  FormGroup,
+  Select,
+  SelectOption,
+  SelectVariant,
+} from "@patternfly/react-core";
+
+import type { ClientQuery } from "@keycloak/keycloak-admin-client/lib/resources/clients";
+import { useAdminClient, useFetch } from "../../context/auth/AdminClient";
+import { HelpItem } from "../help-enabler/HelpItem";
+import type { ComponentProps } from "../dynamic/components";
+
+type ClientSelectProps = ComponentProps & {
+  namespace: string;
+  required?: boolean;
+};
+
+export const ClientSelect = ({
+  name,
+  label,
+  helpText,
+  defaultValue,
+  namespace,
+  isDisabled = false,
+  required = false,
+}: ClientSelectProps) => {
+  const { t } = useTranslation(namespace);
+  const { control, errors } = useFormContext();
+
+  const [open, setOpen] = useState(false);
+  const [clients, setClients] = useState<JSX.Element[]>();
+  const [search, setSearch] = useState("");
+
+  const adminClient = useAdminClient();
+
+  useFetch(
+    () => {
+      const params: ClientQuery = {
+        max: 20,
+      };
+      if (search) {
+        params.clientId = search;
+        params.search = true;
+      }
+      return adminClient.clients.find(params);
+    },
+    (clients) =>
+      setClients(
+        clients.map((option) => (
+          <SelectOption key={option.id} value={option.clientId} />
+        ))
+      ),
+    [search]
+  );
+
+  return (
+    <FormGroup
+      label={t(label!)}
+      isRequired={required}
+      labelIcon={
+        <HelpItem
+          helpText={t(helpText!)}
+          fieldLabelId={`${namespace}:${label}`}
+        />
+      }
+      fieldId={name!}
+      validated={errors[name!] ? "error" : "default"}
+      helperTextInvalid={t("common:required")}
+    >
+      <Controller
+        name={name!}
+        defaultValue={defaultValue || ""}
+        control={control}
+        rules={required ? { required: true } : {}}
+        render={({ onChange, value }) => (
+          <Select
+            toggleId={name}
+            variant={SelectVariant.typeahead}
+            onToggle={(open) => setOpen(open)}
+            isOpen={open}
+            isDisabled={isDisabled}
+            selections={value}
+            onFilter={(_, value) => {
+              setSearch(value);
+              return clients;
+            }}
+            onSelect={(_, value) => {
+              onChange(value.toString());
+              setOpen(false);
+            }}
+            aria-label={t(label!)}
+          >
+            {clients}
+          </Select>
+        )}
+      />
+    </FormGroup>
+  );
+};
